# Remove primitive-projection related data from the kernel