This PR removes redundant or even outright wrong data from data structures going through the kernel.

- `const_proj` from `const_body`: this is redundant with the data stored in the environment since #7234
- `ProjectionEntry` from `constant_entry`: removal of the above fields allow to define projection bodies out of the kernel
- `proj_eta` from `projection_body`: only needed by `ProjectionEntry`, no need to cache it in the kernel
- `proj_body` from `projection_body`: only needed by compatibility printing and extraction.

In the last case, this is completely wrong: the generated term does not even type-check in the kernel due to the fact it uses pattern-matching over negative records.

Several followup PRs are pending on this one, but I'm not publishing them before this one is merged not to flood everybody with PR dependencies.