@maximedenes 2018-06-17
Oh, and btw @ppedrot, I'm afraid this need a rebase because of your other PR fixing a soundness bug. Thanks!