@mattam82 2018-06-17
Looks fine to me. I don’t believe putting the projection compat printing outside changes much performance-wise. I don’t see what would be “retyped” rather than kept here. Do I miss something?