@ppedrot 2018-06-17
@maximedenes Nitpicks addressed, rebased and benchmark running there: https://ci.inria.fr/coq/job/benchmark-part-of-the-branch/451/.