@ppedrot 2018-06-18
The benchmark just shows noise:
```
┌──────────────────────────┬─────────────────────────┬───────────────────────────────────────┬───────────────────────────────────────┬─────────────────────────┬─────────────────┐
│                          │      user time [s]      │              CPU cycles               │           CPU instructions            │  max resident mem [KB]  │   mem faults    │
│                          │                         │                                       │                                       │                         │                 │
│             package_name │     NEW     OLD PDIFF   │            NEW            OLD PDIFF   │            NEW            OLD PDIFF   │     NEW     OLD PDIFF   │ NEW OLD PDIFF   │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-flocq │  260.54  263.31 -1.05 % │   723060309791   730030748144 -0.95 % │   895503065339   899348647831 -0.43 % │ 1245680 1245420 +0.02 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                  coq-vst │ 3750.26 3780.93 -0.81 % │ 10426374020712 10512732716640 -0.82 % │ 13667775890378 13781915333424 -0.83 % │ 2234772 2235312 -0.02 % │   3  41 -92.68 % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-sf-lf │   19.87   20.03 -0.80 % │    53490734453    53539004402 -0.09 % │    69147657547    69273821679 -0.18 % │  418320  418888 -0.14 % │   2   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│              coq-unimath │ 1278.35 1284.55 -0.48 % │  3535428307034  3553268031930 -0.50 % │  6022545976169  6035279703149 -0.21 % │ 1055256 1055448 -0.02 % │   2   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-ssreflect │   45.46   45.66 -0.44 % │   124511104330   124380685099 +0.10 % │   146153618677   146428131121 -0.19 % │  536656  537324 -0.12 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│              coq-bignums │   79.84   80.08 -0.30 % │   219649724906   220627427372 -0.44 % │   283372596766   285977057543 -0.91 % │  521992  522984 -0.19 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                 coq-corn │ 1547.79 1551.34 -0.23 % │  4281498709155  4292684808156 -0.26 % │  6324231343891  6394160082104 -1.09 % │  848996  848872 +0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│    coq-mathcomp-solvable │  201.36  201.48 -0.06 % │   556800918680   557039380836 -0.04 % │   767392798909   768109757355 -0.09 % │  866124  842484 +2.81 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│       coq-mathcomp-field │  450.38  450.63 -0.06 % │  1249058422127  1251209795975 -0.17 % │  2026689912660  2027388378238 -0.03 % │  812952  814316 -0.17 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                 coq-hott │  310.58  310.63 -0.02 % │   863808919817   860168729133 +0.42 % │  1274567014134  1275499779649 -0.07 % │  629560  588540 +6.97 % │  27  26 +3.85 % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│                coq-color │  719.24  718.49 +0.10 % │  1988360164015  1987163533129 +0.06 % │  2358458212165  2369203887740 -0.45 % │ 1414036 1412196 +0.13 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│         coq-math-classes │  226.10  225.82 +0.12 % │   620906599354   618138734452 +0.45 % │   771064730875   771580237407 -0.07 % │  531972  531360 +0.12 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│      coq-formal-topology │   38.05   38.00 +0.13 % │   101562186512   101758322505 -0.19 % │   126533040612   126697453553 -0.13 % │  481900  482276 -0.08 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-character │  273.54  273.06 +0.18 % │   758499527795   757381242907 +0.15 % │  1078277062431  1078899369624 -0.06 % │ 1070672 1072920 -0.21 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│ coq-mathcomp-real-closed │  169.04  168.73 +0.18 % │   468793114098   468413330324 +0.08 % │   684450057392   685371477037 -0.13 % │  837488  838628 -0.14 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│    coq-mathcomp-fingroup │   65.26   65.14 +0.18 % │   179248552259   178917817788 +0.18 % │   238770423028   239143509566 -0.16 % │  592728  593268 -0.09 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│         coq-fiat-parsers │  703.35  701.76 +0.23 % │  1948469449200  1942820291691 +0.29 % │  2971745829163  2973184532351 -0.05 % │ 3377572 3377304 +0.01 % │  26 362 -92.82 % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│     coq-mathcomp-algebra │  184.82  184.34 +0.26 % │   510084253980   509378066693 +0.14 % │   686519037826   686936953092 -0.06 % │  647476  648024 -0.08 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-sf-plf │   52.28   52.13 +0.29 % │   142927874587   142566565498 +0.25 % │   178744742159   178960096646 -0.12 % │  517336  517716 -0.07 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│           coq-coquelicot │   82.36   82.05 +0.38 % │   227273574858   226653206452 +0.27 % │   268805166515   269360721635 -0.21 % │  715744  716436 -0.10 % │  19  83 -77.11 % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-geocoq │ 2823.23 2810.92 +0.44 % │  7827130121038  7795220399087 +0.41 % │ 12500646868881 12503886114078 -0.03 % │ 1388288 1388160 +0.01 % │  64   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│            coq-fiat-core │  112.16  111.61 +0.49 % │   311479707674   310186447996 +0.42 % │   380116784909   380279847723 -0.04 % │  498116  498452 -0.07 % │ 298   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│   coq-mathcomp-odd-order │ 1420.54 1410.18 +0.73 % │  3946359500530  3917729822522 +0.73 % │  6645182689005  6648240256675 -0.05 % │ 1393724 1379208 +1.05 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│               coq-sf-vfa │   25.92   25.68 +0.93 % │    70200392335    70278921782 -0.11 % │    90433365667    90671019924 -0.26 % │  534392  534804 -0.08 % │  22  24 -8.33 % │
├──────────────────────────┼─────────────────────────┼───────────────────────────────────────┼───────────────────────────────────────┼─────────────────────────┼─────────────────┤
│             coq-compcert │  872.20  863.89 +0.96 % │  2415544591512  2393010357461 +0.94 % │  3458737629946  3461280398908 -0.07 % │ 1327332 1327756 -0.03 % │   0   0  +nan % │
└──────────────────────────┴─────────────────────────┴───────────────────────────────────────┴───────────────────────────────────────┴─────────────────────────┴─────────────────┘


PDIFF = proportional difference between measurements done for the NEW and the OLD Coq version
      = (NEW_measurement - OLD_measurement) / OLD_measurement * 100%

NEW = e43710b391c278ac7fcb808ec28d720b4317660c
OLD = d62354b7e9ff8e20aa959984b392a27e26f9fc24
```